Pull OpenShowVisitor out of OpenShow.kt.
